(feat) Implement facebook login/signup in react-native [Closes #370]
- this was achieved with the react-native-facebook-login npm module
and an open PR that has not been merged yet.
- react-native-facebook-login has been added to the tippus package.json
but until the PR is merged it will have to be manually replaced with our
version
- currently uses facebook id which is apparently different for a user
depending on what app the request comes from, so for now the artist accounts
that are created are not linked between webapp and react-native. We
should be able to fix this by using email or another unique id
returned from facebook auth.